Migrate to Isaacsim 5.1.0
Migrated project to support Isaacsim 5.1.0, PyTorch 2.7, TensorDict 0.10.0, and TorchRL 0.10.0. Changes are still preliminary but I have confirmed that PPO runs in all environments as expected. Open to further tests and help with any missing conversions.

The most major breaking change I needed to make was in the Forest environment, where the previously embedded set of composite specs didn't seem to be compatible with the ravel transform anymore, so I had to flatten the lidar into the obs.